+++ src/interfaces/statics.rs
@@ -1,9 +1,5 @@
-use crate::{
-    api::{self, CoverCrypt},
-    cover_crypt_core,
-    error::Error,
-    partitions,
-};
+use std::{ops::DerefMut, sync::Mutex};
+
 use abe_policy::{AccessPolicy, Policy};
 use cosmian_crypto_core::{
     asymmetric_crypto::{curve25519::X25519KeyPair, DhKeyPair},
@@ -11,7 +7,13 @@ use cosmian_crypto_core::{
     symmetric_crypto::{aes_256_gcm_pure::Aes256GcmCrypto, Dem},
     CsRng, KeyTrait,
 };
-use std::{ops::DerefMut, sync::Mutex};
+
+use crate::{
+    api::{self, CoverCrypt},
+    cover_crypt_core,
+    error::Error,
+    partitions,
+};
 
 const TAG_LENGTH: usize = 32;
 
@@ -299,16 +301,23 @@ pub type SymmetricKey = <CoverCryptDem as Dem<{ CoverCryptDem::KEY_LENGTH }>>::K
 
 #[cfg(test)]
 mod tests {
-    use super::*;
-    use crate::{partitions::Partition, CoverCrypt, Error};
     use abe_policy::{AccessPolicy, Attribute, Policy, PolicyAxis};
     use cosmian_crypto_core::bytes_ser_de::{Deserializer, Serializable};
     use serde::{Deserialize, Serialize};
 
+    use super::*;
+    use crate::{partitions::Partition, CoverCrypt, Error};
+
     fn policy() -> Result<Policy, Error> {
         let sec_level = PolicyAxis::new(
             "Security Level",
-            &["Protected", "Confidential", "Top Secret"],
+            &[
+                "Protected",
+                "Low Secret",
+                "Medium Secret",
+                "High Secret",
+                "Top Secret",
+            ],
             true,
         );
         let department = PolicyAxis::new("Department", &["R&D", "HR", "MKG", "FIN"], false);
@@ -339,8 +348,8 @@ mod tests {
         for p in &new_partitions_msk {
             assert!(new_partitions_mpk.contains(p));
         }
-        // 3 is the size of the security level axis
-        assert_eq!(new_partitions_msk.len(), partitions_msk.len() + 3);
+        // 5 is the size of the security level axis
+        assert_eq!(new_partitions_msk.len(), partitions_msk.len() + 5);
         Ok(())
     }
 
@@ -350,7 +359,7 @@ mod tests {
         let cover_crypt = CoverCryptX25519Aes256::default();
         let (mut msk, mut mpk) = cover_crypt.generate_master_keys(&policy)?;
         let access_policy = AccessPolicy::from_boolean_expression(
-            "Department::MKG && Security Level::Confidential",
+            "Department::MKG && Security Level::Medium Secret",
         )?;
         let mut usk = cover_crypt.generate_user_secret_key(&msk, &access_policy, &policy)?;
         let original_usk = usk.clone();
@@ -360,9 +369,9 @@ mod tests {
         cover_crypt.update_master_keys(&policy, &mut msk, &mut mpk)?;
         // refresh the user key and preserve access to old partitions
         cover_crypt.refresh_user_secret_key(&mut usk, &access_policy, &msk, &policy, true)?;
-        // 2 partitions accessed by the user were rotated (MKG Confidential and MKG
+        // 3 partitions accessed by the user were rotated (MKG Medium Secret and MKG
         // Protected)
-        assert_eq!(usk.x.len(), original_usk.x.len() + 2);
+        assert_eq!(usk.x.len(), original_usk.x.len() + 3);
         for x_i in &original_usk.x {
             assert!(usk.x.contains(x_i));
         }
@@ -401,139 +410,261 @@ mod tests {
     }
 
     #[derive(Debug, Serialize, Deserialize)]
+    struct EncryptionTestVector {
+        encryption_policy: String,
+        plaintext: String,
+        ciphertext: String,
+        header_metadata: String,
+        authentication_data: String,
+    }
+
+    impl EncryptionTestVector {
+        fn decrypt(&self, user_key: &str) -> Result<(), Error> {
+            let user_key = UserSecretKey::try_from_bytes(&base64::decode(user_key).unwrap())?;
+
+            let ciphertext = base64::decode(&self.ciphertext).unwrap();
+            let expected_plaintext = base64::decode(&self.plaintext).unwrap();
+
+            let header_metadata = if !self.header_metadata.is_empty() {
+                base64::decode(&self.header_metadata).unwrap()
+            } else {
+                vec![]
+            };
+
+            let authentication_data = if !self.authentication_data.is_empty() {
+                base64::decode(&self.authentication_data).unwrap()
+            } else {
+                vec![]
+            };
+            let authentication_data = if authentication_data.is_empty() {
+                None
+            } else {
+                Some(authentication_data.as_slice())
+            };
+
+            let mut de = Deserializer::new(ciphertext.as_slice());
+            let encrypted_header = EncryptedHeader::read(&mut de)?;
+            let ciphertext = de.finalize();
+            let cover_crypt = CoverCryptX25519Aes256::default();
+
+            let plaintext_header =
+                encrypted_header.decrypt(&cover_crypt, &user_key, authentication_data)?;
+            assert_eq!(plaintext_header.additional_data, header_metadata);
+            let plaintext = cover_crypt.decrypt(
+                &plaintext_header.symmetric_key,
+                &ciphertext,
+                authentication_data,
+            )?;
+            assert_eq!(expected_plaintext, plaintext);
+
+            Ok(())
+        }
+
+        pub fn new(
+            mpk: &PublicKey,
+            policy: &Policy,
+            encryption_policy: &str,
+            plaintext: &str,
+            header_metadata: Option<&[u8]>,
+            authentication_data: Option<&[u8]>,
+        ) -> Result<EncryptionTestVector, Error> {
+            let cover_crypt = CoverCryptX25519Aes256::default();
+            let (symmetric_key, encrypted_header) = EncryptedHeader::generate(
+                &cover_crypt,
+                policy,
+                mpk,
+                &AccessPolicy::from_boolean_expression(encryption_policy)?,
+                header_metadata,
+                authentication_data,
+            )?;
+
+            let mut aes_ciphertext =
+                cover_crypt.encrypt(&symmetric_key, plaintext.as_bytes(), authentication_data)?;
+            let mut encrypted_bytes = encrypted_header.try_to_bytes()?;
+            encrypted_bytes.append(&mut aes_ciphertext);
+            let header_metadata = match header_metadata {
+                Some(ad) => base64::encode(ad),
+                None => String::new(),
+            };
+            let authentication_data = match authentication_data {
+                Some(ad) => base64::encode(ad),
+                None => String::new(),
+            };
+            Ok(EncryptionTestVector {
+                encryption_policy: encryption_policy.to_string(),
+                plaintext: base64::encode(plaintext),
+                ciphertext: base64::encode(encrypted_bytes),
+                header_metadata,
+                authentication_data,
+            })
+        }
+    }
+
+    #[derive(Debug, Serialize, Deserialize)]
+    struct UserSecretKeyTestVector {
+        access_policy: String,
+        key: String,
+    }
+    impl UserSecretKeyTestVector {
+        pub fn new(
+            msk: &MasterSecretKey,
+            policy: &Policy,
+            access_policy: &str,
+        ) -> Result<Self, Error> {
+            Ok(Self {
+                key: base64::encode(
+                    CoverCryptX25519Aes256::default()
+                        .generate_user_secret_key(
+                            msk,
+                            &AccessPolicy::from_boolean_expression(access_policy)?,
+                            policy,
+                        )?
+                        .try_to_bytes()?,
+                ),
+                access_policy: access_policy.to_string(),
+            })
+        }
+    }
+
+    #[derive(Debug, Serialize, Deserialize)]
     struct NonRegressionTestVector {
         public_key: String,
-        msk: String,
+        master_secret_key: String,
         policy: String,
-        user_decryption_key: String,
-        user_decryption_key_2: String,
-        encrypted_bytes: String,
-        plaintext: String,
-        additional_data: String,
+        top_secret_mkg_fin_key: UserSecretKeyTestVector,
+        medium_secret_mkg_key: UserSecretKeyTestVector,
+        top_secret_fin_key: UserSecretKeyTestVector,
+        low_secret_mkg_test_vector: EncryptionTestVector,
+        top_secret_mkg_test_vector: EncryptionTestVector,
+        low_secret_fin_test_vector: EncryptionTestVector,
+    }
+
+    impl NonRegressionTestVector {
+        pub fn new() -> Result<Self, Error> {
+            //
+            // Policy settings
+            //
+            let mut policy = policy()?;
+            policy.rotate(&Attribute::new("Department", "FIN"))?;
+
+            //
+            // CoverCrypt setup
+            //
+            let cover_crypt = CoverCryptX25519Aes256::default();
+            let (msk, mpk) = cover_crypt.generate_master_keys(&policy)?;
+
+            //
+            // Encryption header metadata
+            let header_metadata = 1u32.to_be_bytes().to_vec();
+            let authentication_data = 2u32.to_be_bytes().to_vec();
+
+            let reg_vectors = NonRegressionTestVector {
+                public_key: base64::encode(mpk.try_to_bytes()?),
+                master_secret_key: base64::encode(msk.try_to_bytes()?),
+                policy: base64::encode(serde_json::to_vec(&policy)?),
+                //
+                // Create user decryption keys
+                top_secret_mkg_fin_key: UserSecretKeyTestVector::new(
+                    &msk,
+                    &policy,
+                    "(Department::MKG || Department:: FIN) && Security Level::Top Secret",
+                )?,
+                medium_secret_mkg_key: UserSecretKeyTestVector::new(
+                    &msk,
+                    &policy,
+                    "Security Level::Medium Secret && Department::MKG",
+                )?,
+                top_secret_fin_key: UserSecretKeyTestVector::new(
+                    &msk,
+                    &policy,
+                    "Security Level::Top Secret && Department::FIN",
+                )?,
+                //
+                // Generate ciphertexts
+                top_secret_mkg_test_vector: EncryptionTestVector::new(
+                    &mpk,
+                    &policy,
+                    "Department::MKG && Security Level::Top Secret",
+                    "top_secret_mkg_plaintext",
+                    Some(&header_metadata),
+                    Some(&authentication_data),
+                )?,
+
+                low_secret_mkg_test_vector: EncryptionTestVector::new(
+                    &mpk,
+                    &policy,
+                    "Department::MKG && Security Level::Low Secret",
+                    "low_secret_mkg_plaintext",
+                    Some(&header_metadata),
+                    None,
+                )?,
+
+                low_secret_fin_test_vector: EncryptionTestVector::new(
+                    &mpk,
+                    &policy,
+                    "Department::FIN && Security Level::Low Secret",
+                    "low_secret_fin_plaintext",
+                    None,
+                    None,
+                )?,
+            };
+            Ok(reg_vectors)
+        }
+
+        fn verify(&self) -> Result<(), Error> {
+            // top_secret_fin_key
+            self.low_secret_fin_test_vector
+                .decrypt(&self.top_secret_fin_key.key)?;
+            assert!(self
+                .low_secret_mkg_test_vector
+                .decrypt(&self.top_secret_fin_key.key)
+                .is_err());
+            assert!(self
+                .top_secret_mkg_test_vector
+                .decrypt(&self.top_secret_fin_key.key)
+                .is_err());
+
+            // top_secret_mkg_fin_key
+            self.low_secret_fin_test_vector
+                .decrypt(&self.top_secret_mkg_fin_key.key)?;
+            self.low_secret_mkg_test_vector
+                .decrypt(&self.top_secret_mkg_fin_key.key)?;
+            self.top_secret_mkg_test_vector
+                .decrypt(&self.top_secret_mkg_fin_key.key)?;
+
+            // medium_secret_mkg_key
+            assert!(self
+                .low_secret_fin_test_vector
+                .decrypt(&self.medium_secret_mkg_key.key)
+                .is_err());
+            self.low_secret_mkg_test_vector
+                .decrypt(&self.medium_secret_mkg_key.key)?;
+            assert!(self
+                .top_secret_mkg_test_vector
+                .decrypt(&self.medium_secret_mkg_key.key)
+                .is_err());
+            Ok(())
+        }
     }
 
     #[test]
-    fn test_hybrid_encryption_decryption() -> Result<(), Error> {
-        //
-        // Policy settings
-        //
-        let sec_level = PolicyAxis::new(
-            "Security Level",
-            &[
-                "Protected",
-                "Low Secret",
-                "Medium Secret",
-                "High Secret",
-                "Top Secret",
-            ],
-            true,
-        );
-        let department = PolicyAxis::new("Department", &["R&D", "HR", "MKG", "FIN"], false);
-        let mut policy = Policy::new(100);
-        policy.add_axis(&sec_level)?;
-        policy.add_axis(&department)?;
-        policy.rotate(&Attribute::new("Department", "FIN"))?;
-        let access_policy = AccessPolicy::from_boolean_expression(
-            "(Department::HR || Department:: FIN) && Security Level::Low Secret",
-        )?;
-
-        //
-        // CoverCrypt setup
-        //
-        let cover_crypt = CoverCryptX25519Aes256::default();
-        let (msk, mpk) = cover_crypt.generate_master_keys(&policy)?;
-        let top_secret_mkg_fin_user =
-            cover_crypt.generate_user_secret_key(&msk, &access_policy, &policy)?;
-
-        //
-        // Encrypt/decrypt header
-        //
-        let additional_data = 1u32.to_be_bytes().to_vec();
-        let authentication_data = None;
-
-        let (symmetric_key, encrypted_header) = EncryptedHeader::generate(
-            &cover_crypt,
-            &policy,
-            &mpk,
-            &access_policy,
-            Some(&additional_data),
-            authentication_data,
-        )?;
-        let res = encrypted_header.decrypt(
-            &cover_crypt,
-            &top_secret_mkg_fin_user,
-            authentication_data,
-        )?;
-
-        assert_eq!(additional_data, res.additional_data);
-
-        let message = b"My secret message";
-        // we need mut in the commented lines below
-        #[allow(unused_mut)]
-        let mut ctx = cover_crypt.encrypt(&symmetric_key, message, authentication_data)?;
-
-        let res = cover_crypt.decrypt(&symmetric_key, &ctx, authentication_data)?;
-
-        assert_eq!(message.to_vec(), res);
-
-        //
-        // Uncomment the following code to write a new regression vector
-        //
-
-        //let mut encrypted_bytes = encrypted_header.try_to_bytes()?;
-        //encrypted_bytes.append(&mut ctx);
-
-        //let access_policy_2 = AccessPolicy::new("Security Level", "Medium Secret")
-        //& AccessPolicy::new("Department", "MKG");
-
-        //let medium_secret_mkg_user =
-        //cover_crypt.generate_user_secret_key(&msk, &access_policy_2, &policy)?;
-
-        //let reg_vectors = NonRegressionTestVector {
-        //public_key: hex::encode(mpk.try_to_bytes()?),
-        //msk: hex::encode(msk.try_to_bytes()?),
-        //policy: hex::encode(serde_json::to_vec(&policy)?),
-        //user_decryption_key: hex::encode(top_secret_mkg_fin_user.try_to_bytes()?),
-        //user_decryption_key_2: hex::encode(medium_secret_mkg_user.try_to_bytes()?),
-        //encrypted_bytes: hex::encode(encrypted_bytes),
-        //plaintext: hex::encode(message),
-        //additional_data: hex::encode(additional_data),
-        //};
-
-        //std::fs::write(
-        //"non_regression_vector.json",
-        //serde_json::to_string(&reg_vectors).unwrap(),
-        //)
-        //.unwrap();
+    fn test_generate_non_regression_vector() -> Result<(), Error> {
+        let _reg_vector = NonRegressionTestVector::new()?;
+        // uncomment this to regenerate new test vector
+        // std::fs::write(
+        //     "tests_data/non_regression_vector.json",
+        //     serde_json::to_string(&reg_vector).unwrap(),
+        // )
+        // .unwrap();
 
         Ok(())
     }
 
     #[test]
-    fn test_non_regregression() {
+    fn test_non_regression() -> Result<(), Error> {
         let reg_vector: NonRegressionTestVector =
-            serde_json::from_str(include_str!("../../non_regression_vector.json")).unwrap();
-        let top_secret_mkg_fin_user =
-            UserSecretKey::try_from_bytes(&hex::decode(reg_vector.user_decryption_key).unwrap())
-                .unwrap();
-        let medium_secret_mkg_user =
-            UserSecretKey::try_from_bytes(&hex::decode(reg_vector.user_decryption_key_2).unwrap())
-                .unwrap();
-        let encrypted_bytes = &hex::decode(reg_vector.encrypted_bytes).unwrap();
-        let mut de = Deserializer::new(encrypted_bytes.as_slice());
-        let encrypted_header = EncryptedHeader::read(&mut de).unwrap();
-        let ciphertext = de.finalize();
-        let cover_crypt = CoverCryptX25519Aes256::default();
-        assert!(encrypted_header
-            .decrypt(&cover_crypt, &medium_secret_mkg_user, None)
-            .is_err());
-        let cleartext_header = encrypted_header
-            .decrypt(&cover_crypt, &top_secret_mkg_fin_user, None)
-            .unwrap();
-        let plaintext = cover_crypt
-            .decrypt(&cleartext_header.symmetric_key, &ciphertext, None)
-            .unwrap();
-        assert_eq!(hex::decode(&reg_vector.plaintext).unwrap(), plaintext);
+            serde_json::from_str(include_str!("../../tests_data/non_regression_vector.json"))?;
+        reg_vector.verify()
     }
 
     #[test]
@@ -591,7 +722,7 @@ mod tests {
             None,
         )?;
 
-        let _cleartext_header =
+        let _plaintext_header =
             encrypted_header.decrypt(&cover_crypt, &top_secret_fin_usk, None)?;
 
         //