+++ CHANGELOG.md
@@ -8,6 +8,8 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
+- non regression test vectors with different encryption policies
+
 ### Changed
 
 - improve serialization